### 2022-06-09 #4422 @hrossman
I would love to have @CamDavidsonPilon, author of the awesome lifelines package (which also serves our package very well). I know he has published on JOSS but do not know his availability for review.
Thank you for your time and effort on this!
