### 2022-06-09 #4422 @majensen
Sounds like a plan: @CamDavidsonPilon, would you be willing to review [this work](#4422) for [JOSS](https://joss.theoj.org)? Our guide for reviewers is at https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. We would appreciate your expertise very much - thanks for your consideration - --@majensen
