### 2022-06-10 33546d6
Update methods.md
