### 2022-06-10 5ace3d0
Update methods.md
