### 2022-06-10 8ce04f6
Update methods.md
