### 2022-06-15 hrossman/pymsm#56 @CamDavidsonPilon
Automated tests
👋 Hi folks, I'm the reviewer for your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4422). I'll be making issues in this repo for my feedback.

-----

1. Currently, github actions will only run automated tests (CI) on [pull requests](https://github.com/hrossman/pymsm/blob/main/.github/workflows/tests.yml#L3). I would suggest making this on `pull_request` _and_ `push`. Contributors will push to the main branch, and you want CI to run for these commits, too.
2. There is installation instructions to install from pypi, but what about locally? I git pulled the repo, ran `pip3 install -e .`, and the `py.test` to test. That can be documented in a section on the README or the docs.
3. tests in `test_msm_examples.py` and `test_sim.py` should assert some output. There should be a reason you are running a test - assert that.
4. This can be fixed (found running py.test on my machine, py3.9.6)
```
src/pymsm/datasets/__init__.py:319
  /Users/camerondavidson-pilon/code/pymsm/src/pymsm/datasets/__init__.py:319: DeprecationWarning: invalid escape sequence \R
    1: "Discharged\Recovered",
```
5. If possible, try to fix the lifelines warnings that show up. Not always possible, I know.
6. `stepfunc` seems like an important function, but has lots of potential edge cases. This would be a good function to test as well.
