### 2022-06-15 #4422 @hrossman
Thank you for the detailed review @CamDavidsonPilon , Much appreciated!
We will start work to address all of the issues you raised

@majensen  - The plan is to address issues & push all relevant code, docs, and paper changes. After that, we will update and summarize changes here at the end. is that Ok or is there any other prefered process?
Thanks both
