### 2022-06-16 6dd08e7
Create CONTRIBUTING.md
