+++ README.md
@@ -1,5 +1,5 @@
-![Docs](https://github.com/hrossman/pymsm/actions/workflows/docs.yml/badge.svg)
-![Tests](https://github.com/hrossman/pymsm/actions/workflows/tests.yml/badge.svg)
+[![Docs](https://github.com/hrossman/pymsm/actions/workflows/docs.yml/badge.svg)](https://hrossman.github.io/pymsm/)
+[![Tests](https://github.com/hrossman/pymsm/actions/workflows/tests.yml/badge.svg)](https://github.com/hrossman/pymsm/actions/workflows/tests.yml)
 [![codecov](https://codecov.io/gh/hrossman/pymsm/branch/main/graph/badge.svg?token=FG434UHSQ2)](https://codecov.io/gh/hrossman/pymsm)
 [![pypi version](https://img.shields.io/pypi/v/pymsm)](https://pypi.org/project/pymsm/)
 [![DOI](https://zenodo.org/badge/443028256.svg)](https://zenodo.org/badge/latestdoi/443028256)
