### 2022-06-28 #4422 @majensen
@tedinburgh would you have time and willingness to review [this work](#4422) for [JOSS](https://joss.theoj.org)? I would appreciate your expertise. Please have a look at our [reviewer guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and let me know in this thread. Thanks for your consideration! --@majensen
