### 2022-06-30 #4422 @majensen
@becarioprecario would you be available and willing to review [this work](#4422) for [JOSS](https://joss.theoj.org)? Your survival analysis expertise would be very helpful. Thanks for your consideration --@majensen
