### 2022-07-12 #4422 @majensen
@stefanocorreta - would you be able to review [this work](#4422) for [JOSS](https://joss.theoj.org) instead? Really appreciate your consideration! --@majensen
