### 2022-07-12 #4422 @majensen
@stefanocoretta - would you be able to review [this work](https://github.com/openjournals/joss-reviews/issues/4422#4422) for [JOSS](https://joss.theoj.org/) instead? Really appreciate your consideration! --@majensen
