### 2022-07-12 #4422 @becarioprecario
Thanks for the invitation but I am afraid that I am not available at the moment.

El 30 jun 2022, a las 21:41, Mark Jensen ***@***.******@***.***>> escribió:

@becarioprecario<https://github.com/becarioprecario> would you be available and willing to review this work<x-msg://1/#4422> for JOSS<https://joss.theoj.org/>? Your survival analysis expertise would be very helpful. Thanks for your consideration ***@***.***<https://github.com/majensen>
