### 2022-07-12 #4566 @editorialbot
[REVIEW]: PyMSM: Python package for Competing Risks and Multi-State models for Survival Data
**Submitting author:** @hrossman (<a href="http://orcid.org/0000-0002-5613-8004">Hagai Rossman</a>)
**Repository:** https://github.com/hrossman/pymsm
**Branch with paper.md** (empty if default branch):
**Version:** v0.2
