+++ mkdocs.yml
@@ -81,7 +81,6 @@ nav:
       - api/multistate_model.md
       - api/multistate_simulator.md
       - api/pathobject.md
-    - methods.md
 
 
 plugins:
