### 2022-07-25 #4566 @majensen
@stefanocoretta yes - it is in the "paper.md" file at the top level of the code repo (https://github.com/hrossman/pymsm/blob/main/paper.md). The bibliography is in "paper.bib".
