### 2022-08-04 hrossman/pymsm#56 @CamDavidsonPilon
@hrossman what are your thoughts on the above? Any confusion about the provided feedback?
