### 2022-08-04 hrossman/pymsm#56 @hrossman
Agree with all comments. Was waiting for all reviewers before we deal with all comments jointly. Thank you, we'll start working on these
