### 2022-08-04 #4566 @hrossman
Hi all,
So if I understand correctly, we are done with this round of review?
We will start fixing all issues that I am aware of and ping you here once done.
I apologize if some have been hanging for some time @CamDavidsonPilon, I was waiting to go through and fix the comments from both reviewers jointly and have had some busy schedule.

@stefanocoretta - just to make sure - your comments are in the Github repo issues right? And the checklist here?

Thank you all for your prompt effort
