### 2022-08-28 590a75c
add local installation to readme
