### 2022-08-30 cbc4ca1
Add stepfunc test
