### 2022-08-30 8920ad5
add assertions to tests
