### 2022-08-31 bc9b668
tests and rerun notebooks
