### 2022-08-31 hrossman/pymsm#56 @hrossman
Issues status:
1. Fixed
2. Added local installation instructions to README
3. Added assertions
4. Fixed
5. Noted thanks. The one place they do show up is in the COVID example notebook, wheere we actually want to to show users how to deal with such a situation. We comment on that in the notebook as a precursor to the "transition trimming" procedure.
6. Indeed thanks. Added tests for `stepfunc` function now.
