### 2022-09-01 #4566 @hrossman
Hi all,
Ok, so I see @CamDavidsonPilon Review checklist is complete
regarding open points in @stefanocoretta review checklist:

1 -  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Reply: Installation is properly managed via Pypi. Instructions available both for pip installation and local git clone installation in the README and docs. Dependencies managed automatically through the pyproject.toml file and pip.

2 -

Reply: Yes, see updated [CONTRIBUTING.md](https://github.com/hrossman/pymsm/blob/main/CONTRIBUTING.md)

3 - State of the field: Do the authors describe how this software compares to other commonly-used packages?
We believe we give the background and motivation and then expliciltly state in the paper (section "Statement of Need"): "To the authors best knowledge, this is the first open-source multi-state model tool that allows fitting of such models while also dealing with important concepts such as right censoring, competing events, recurrent events, left truncation, and time-dependent covariates."

Thanks for all the great comments! Plz let us know if we are in good enough shape, or would you like anything else taken care of?
@majensen
