### 2022-09-03 #4566 @majensen
Thanks @hrossman - I think we can head towards the exit! I will take an editorial look at the paper itself, this might generate some (likely minor) comments. I'll put those here soon. Once the paper text is finalized, I'll give some info about creating an open archive of the code and proceeding to the final sign off by an associate editor-in-chief.
