### 2022-10-02 #4566 @majensen
@hrossman - thanks very much for this submission, and for your patience. The package provides a number of very useful features based on the estimators you have developed.

I have just a couple of wording suggestions, below. The line numbers are based on the PDF proof generated at https://github.com/openjournals/joss-reviews/issues/4566#issuecomment-1236207482 above. I would also look at the bot-suggested references (https://github.com/openjournals/joss-reviews/issues/4566#issuecomment-1264700876) “MISSING DOIs” and see if these are appropriate to incorporate into the bibilography.

My main thought is, can we reduce the relative amount of text in order to highlight the application itself, rather than the statistical machinery behind it? I believe the Roimi 2021 paper does a good job of providing most or all of the theoretical detail, so I am suggesting to summarize the steps very briefly and refer back to that work for those details. I have made an attempt at some summaries below; please feel free to use these if they make sense.

Summary: lines 9, 16, 19, 21 could -> can
line 36: could -> can
line 83: implemented -> subclassed
line 96: ‘would be’ -> is; ‘certain’ -> given

- insert after line 104:
“The conditional probability of transition from one given state to another by time t is the hazard function of that state, integrated over the joint differential hazard function in time. The time-dependent conditional probability thus obtained for each possible transition may then be used in a straightforward way to calculate the transition probability between any pair of states for any time interval [t’, t]. Details may be found in the Supplementary Material of Roimi et al 2021.”
- delete rest of section

- line 144: ‘above’ -> ‘present in the transition probability expression described above’
- delete 146-148, add:
“This procedure yields an estimator dependent only on observed data, by means of the regression performed above. Details may be found in the Supplementary Material of Roimi et al 2021.”

- insert after line 160:
“We may use the estimation technique described above to separably predict the state changes and the dwell times in each state for a sojourn. States are sampled using the estimated discrete transition probabilities. Dwell times are sampled by setting the relevant cumulative transition probability estimator to a sample drawn from Uniform[0,1] and solving for time. Details may be found in the Supplementary Material of Roimi et al 2021.”
- delete 161-169.
