### 2022-10-03 #4566 @hrossman
@majensen,  Thank you for the detailed review!
We will take care of all the typos and DOIs shortly.

I’ve consulted with my co-authors and would like to respond to the main point raised before proceeding:
We do believe it important to collate the statistical machinery in this paper as currently presented. We made considerable effort to summarize the important parts of it and believe it is valuable to have it here, and not only in an appendix of a different paper with a specific application (COVID). After trimming and thinking how to minimize the technical parts, we finalized the text such that a reader that will only want to read about the applications can skip reading from row 93 onward. We believe that many readers will want to understand the machinery and the details behind the scenes of these models, and that having all summarized in one place is very helpful.

If you could reconsider and accept our preference, we would be delighted and think that this would be the fullest and best version of the paper we hope to publish to a diverse community of potential readers\users. If not, we will (unhappily) drop and shorten the paper as you requested.
Please let me know what your decision is.
Both ways, thank you again for your time and effort!

Hagai
