### 2022-10-03 #4566 @majensen
Hi @hrossman - I agree with your considerations and we can go ahead with the version as written. Then I would just make the following minor smoothing edits:

105: “T denotes” -> “T denote”
123: “All of the above” -> “For all of the above”
145: “any unknown parameter is replaced by its estimator” -> “replacing any unknown parameter by its estimator”

thanks!
