+++ paper.bib
@@ -51,6 +51,7 @@
 }
 
 @article{Andersen:1982,
+  doi = {10.1214/aos/1176345976},
   url = {https://www.jstor.org/stable/2240714},
   year = {1982},
   month = dec,
@@ -63,6 +64,7 @@
 }
 
 @article{Klein:2006,
+  doi = {10.1111/j.1541-0420.2006.00589_9.x},
   url = {https://link.springer.com/book/10.1007/b97377},
   year = {2006},
   publisher = {Springer Science & Business Media.},
