### 2022-10-03 #4566 @majensen
Thanks @hrossman - now we need to create an open archive of the software repo. Can you please use Zenodo, FigShare or similar to do this? (There is a short tutorial at [this comment](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729).)

Once that is done, please report the assigned DOI back to this thread.
