### 2022-10-04 #4566 @majensen
Thanks @hrossman - would you be able to set the title of the archive to the the title of the paper, and add your third author to the archive authors? This is just a best practice that JOSS prefers. Let me know if that works for you.
