### 2022-10-09 #4566 @hrossman
> @CamDavidsonPilon, @stefanocoretta – many thanks for your reviews here and to @majensen  for editing this submission! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
>
> @hrossman  – your paper is now accepted and published in JOSS :zap::rocket::boom:

Much appreciated @CamDavidsonPilon @stefanocoretta @majensen
Thank you for you the review and effort!
